Adjust the folders
