+++ src/Microsoft.OData.Edm/Csdl/Json/Ast/CsdlJsonSchemaStructuredItem.cs
@@ -1,0 +1,18 @@
+﻿//---------------------------------------------------------------------
+// <copyright file="CsdlJsonSchemaStructuredItem.cs" company="Microsoft">
+//      Copyright (C) Microsoft Corporation. All rights reserved. See License.txt in the project root for license information.
+// </copyright>
+//---------------------------------------------------------------------
+
+namespace Microsoft.OData.Edm.Csdl.Json.Ast
+{
+    internal abstract class CsdlJsonSchemaStructuredItem : CsdlJsonSchemaItem
+    {
+        public string BaseType { get; set; }
+
+        public bool IsAbstract { get; set; }
+
+        public bool IsOpen { get; set; }
+    }
+
+}
