+++ src/Microsoft.OData.Edm/Csdl/Json/Parser/CsdlJsonModelParser.cs
@@ -1,5 +1,5 @@
 ﻿//---------------------------------------------------------------------
-// <copyright file="CsdlJsonParser.cs" company="Microsoft">
+// <copyright file="CsdlJsonModelBuilder.cs" company="Microsoft">
 //      Copyright (C) Microsoft Corporation. All rights reserved. See License.txt in the project root for license information.
 // </copyright>
 //---------------------------------------------------------------------
@@ -7,45 +7,69 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using Microsoft.OData.Edm.Csdl.Json;
+using System.IO;
+using System.Linq;
+using Microsoft.OData.Edm.Csdl.Json.Ast;
+using Microsoft.OData.Edm.Csdl.Json.Reader;
 using Microsoft.OData.Edm.Csdl.Json.Value;
-using Microsoft.OData.Edm.Csdl.Parsing.Ast;
 
-namespace Microsoft.OData.Edm.Csdl.Parsing
+namespace Microsoft.OData.Edm.Csdl.Json.Parser
 {
     /// <summary>
-    /// Provides CSDL-JSON parsing services for EDM models.
+    /// For example we have A, B, C, D models
+    /// A -> B
+    ///      B -> C
+    ///          C -> D
+    ///               D -> B
+    ///      B -> D
+    /// So, If we load start from A, A is the main model, B, C, D are the refereneced models of A
+    ///     If we load start from C, C is the mainn model, D, B are the referenced models of C
     /// </summary>
-    internal static class CsdlJsonParser
+    internal class CsdlJsonModelBuilder
     {
+        private CsdlSerializerOptions _options;
+        private CsdlJsonModel _mainModel;
+        private TextReader _textReader;
+        private string _source;
+        public CsdlJsonModelBuilder(TextReader textReader, CsdlSerializerOptions options)
+            : this(textReader, options, null, null)
+        {
+        }
+
+        public CsdlJsonModelBuilder(TextReader textReader, CsdlSerializerOptions options, string source, CsdlJsonModel mainModel)
+        {
+            _textReader = textReader;
+            _options = options;
+            _source = source;
+            _mainModel = mainModel;
+        }
+
         /// <summary>
-        /// Parse the JSON value <see cref="IJsonValue"/> to <see cref="CsdlModel"/>.
+        /// Parse the JSON value <see cref="IJsonValue"/> to <see cref="CsdlJsonModel"/>.
         /// </summary>
         /// <param name="jsonValue">The JSON value to parse.</param>
         /// <param name="jsonPath">The JSON path to track on the path of the processed JSON value.</param>
         /// <param name="options">The JSON serializer options.</param>
-        /// <returns>Null or the generated <see cref="CsdlModel"/>.</returns>
-        internal static CsdlModel ParseCsdlModel(IJsonValue jsonValue, JsonPath jsonPath, CsdlSerializerOptions options)
+        /// <returns>Null or the generated <see cref="CsdlJsonModel"/>.</returns>
+        internal CsdlJsonModel TryParseCsdlJsonModel()
         {
+            JsonReader jsonReader = new JsonReader(_textReader, _options.GetJsonReaderOptions());
+            JsonObjectValue csdlObject = jsonReader.ReadAsObject();
+
             // A CSDL JSON document consists of a single JSON object.
-            JsonObjectValue csdlObject = jsonValue.ValidateRequiredJsonValue<JsonObjectValue>(jsonPath);
+            // JsonObjectValue csdlObject = jsonValue.ValidateRequiredJsonValue<JsonObjectValue>(jsonPath);
 
-            // This document object MUST contain the member $Version.
-            Version version = GetCsdlVersion(csdlObject, jsonPath);
-            Debug.Assert(version != null);
-
-            CsdlModel csdlModel = new CsdlModel
-            {
-                Version = version
-            };
-
+            IJsonPath jsonPath = new JsonPath(_source);
             IList<IEdmReference> references = null;
+            Version version = null;
+            IDictionary<string, IJsonValue> members = new Dictionary<string, IJsonValue>();
             csdlObject.ProcessProperty(jsonPath, (propertyName, propertyValue) =>
             {
                 switch (propertyName)
                 {
                     case "$Version":
-                        // Processed above, so skip it.
+                        // This document object MUST contain the member $Version.
+                        version = ParseVersion(propertyValue, jsonPath);
                         break;
 
                     case "$EntityContainer":
@@ -56,31 +80,87 @@ namespace Microsoft.OData.Edm.Csdl.Parsing
 
                     case "$Reference":
                         // The document object MAY contain the member $Reference to reference other CSDL documents.
-                        references = ParseReferences(propertyValue, jsonPath, options);
+                        references = ParseReferences(propertyValue, jsonPath);
                         break;
 
                     default:
                         // CSDL document also MAY contain members for schemas.
                         // Each schema's value is an object.
-                        CsdlSchema csdlSchema = SchemaJsonParser.ParseCsdlSchema(propertyName, propertyValue, jsonPath, version, options);
-                        if (csdlSchema != null)
-                        {
-                            csdlModel.AddSchema(csdlSchema);
-                        }
-                        else
-                        {
-                            propertyValue.ReportUnknownMember(jsonPath, options);
-                        }
+                        members[propertyName] = propertyValue;
                         break;
                 }
             });
 
-            if (references != null && references.Count > 0)
+            Debug.Assert(version != null);
+
+            CsdlJsonModel model = new CsdlJsonModel(_source, version);
+            model.AddCurrentModelReferences(references);
+
+            foreach (var member in members)
             {
-                csdlModel.AddCurrentModelReferences(references);
+                CsdlJsonSchemaParser schemaJsonParser = new CsdlJsonSchemaParser(version, member.Key, _options);
+                //schemaJsonParser.TryParseCsdlSchema(member.Value, jsonPath);
+
+                //model.AddSchemaJsonItems(schemaJsonParser.SchemaItems);
+                CsdlJsonSchema schema = schemaJsonParser.TryParseCsdlJsonSchema(member.Value, jsonPath);
+
+                model.AddSchema(schema);
             }
 
-            return csdlModel;
+            model.BuildNamespaceAlias();
+
+            // We are building the reference models, add into the main model.
+            if (_mainModel != null)
+            {
+                _mainModel.AddReferencedModel(model);
+                BuildReferencedModels(_mainModel, references);
+            }
+            else
+            {
+                BuildReferencedModels(model, references);
+            }
+
+            return model;
+        }
+
+        internal void BuildReferencedModels(CsdlJsonModel mainModel, IList<IEdmReference> edmReferences)
+        {
+            if (edmReferences == null)
+            {
+                return;
+            }
+
+            foreach (var edmReference in edmReferences)
+            {
+                // If added, skip to parse it.
+                string sourceUri = edmReference.Uri.OriginalString;
+                if (mainModel.IsReferencedModelAdded(sourceUri))
+                {
+                    continue;
+                }
+
+                // If nothing included, why does it exist?
+                if (!edmReference.Includes.Any() && !edmReference.IncludeAnnotations.Any())
+                {
+                    continue;
+                }
+
+                // Customer can provide their own built-in vocabulary model
+                TextReader referencedJsonReader = _options.ReferencedModelJsonFactory(edmReference.Uri);
+                if (referencedJsonReader == null)
+                {
+                    // If provided, we doesn't need to load it again.
+                    referencedJsonReader = LoadBuiltInVocabulary(edmReference.Uri);
+                }
+
+                if (referencedJsonReader == null)
+                {
+                    throw new Exception();
+                }
+
+                CsdlJsonModelBuilder builder = new CsdlJsonModelBuilder(referencedJsonReader, _options, sourceUri, mainModel);
+                builder.TryParseCsdlJsonModel();
+            }
         }
 
         /// <summary>
@@ -92,7 +172,7 @@ namespace Microsoft.OData.Edm.Csdl.Parsing
         /// <param name="jsonValue"></param>
         /// <param name="jsonPath"></param>
         /// <returns></returns>
-        public static IList<IEdmReference> ParseReferences(IJsonValue jsonValue, JsonPath jsonPath, CsdlSerializerOptions options)
+        public IList<IEdmReference> ParseReferences(IJsonValue jsonValue, IJsonPath jsonPath)
         {
             // The value of $Reference is an object that contains one member per referenced CSDL document.
             JsonObjectValue referenceObj = jsonValue.ValidateRequiredJsonValue<JsonObjectValue>(jsonPath);
@@ -103,14 +183,14 @@ namespace Microsoft.OData.Edm.Csdl.Parsing
             {
                 // The name of the pair is a URI for the referenced document.
                 // The URI MAY be relative to the document containing the $Reference.
-                IEdmReference result = ParseReferenceObject(propertyName, propertyValue, jsonPath, options);
+                IEdmReference result = ParseReferenceObject(propertyName, propertyValue, jsonPath);
                 references.Add(result);
             });
 
             return references;
         }
 
-        public static IEdmReference ParseReferenceObject(string url, IJsonValue jsonValue, JsonPath jsonPath, CsdlSerializerOptions options)
+        public IEdmReference ParseReferenceObject(string url, IJsonValue jsonValue, IJsonPath jsonPath)
         {
             // The value of each reference object is an object.
             JsonObjectValue referenceObj = jsonValue.ValidateRequiredJsonValue<JsonObjectValue>(jsonPath);
@@ -126,20 +206,20 @@ namespace Microsoft.OData.Edm.Csdl.Parsing
                     case "$Include":
                         // The value of $Include is an array.
                         // Array items are objects that MUST contain the member $Namespace and MAY contain the member $Alias.
-                        includes = propertyValue.ParseArray(jsonPath, options, ParseInclude);
+                        includes = propertyValue.ParseArray(jsonPath, ParseInclude);
                         break;
 
                     case "$IncludeAnnotations":
                         // The value of $IncludeAnnotations is an array.
                         // Array items are objects that MUST contain the member $TermNamespace and MAY contain the members $Qualifier and $TargetNamespace.
-                        includeAnnotations = propertyValue.ParseArray(jsonPath, options, ParseIncludeAnnotations);
+                        includeAnnotations = propertyValue.ParseArray(jsonPath, ParseIncludeAnnotations);
                         break;
 
                     default:
                         // The reference objects MAY contain annotations.However, EdmReference doesn't support annotation.
                         // So, skip the annotation.
 
-                        propertyValue.ReportUnknownMember(jsonPath, options);
+                        propertyValue.ReportUnknownMember(jsonPath, _options);
                         break;
                 }
             });
@@ -157,7 +237,7 @@ namespace Microsoft.OData.Edm.Csdl.Parsing
         /// <param name="jsonPath"></param>
         /// <param name="options"></param>
         /// <returns></returns>
-        public static IEdmInclude ParseInclude(IJsonValue jsonValue, JsonPath jsonPath, CsdlSerializerOptions options)
+        public IEdmInclude ParseInclude(IJsonValue jsonValue, IJsonPath jsonPath)
         {
             JsonObjectValue includeObj = jsonValue.ValidateRequiredJsonValue<JsonObjectValue>(jsonPath);
 
@@ -180,7 +260,7 @@ namespace Microsoft.OData.Edm.Csdl.Parsing
 
                     default:
                         // The item objects MAY contain annotations. However, EdmInclude does not supported yet. So skip
-                        propertyValue.ReportUnknownMember(jsonPath, options);
+                        propertyValue.ReportUnknownMember(jsonPath, _options);
                         break;
                 }
             });
@@ -195,7 +275,7 @@ namespace Microsoft.OData.Edm.Csdl.Parsing
         /// <param name="jsonPath"></param>
         /// <param name="options"></param>
         /// <returns></returns>
-        public static IEdmIncludeAnnotations ParseIncludeAnnotations(IJsonValue jsonValue, JsonPath jsonPath, CsdlSerializerOptions options)
+        public IEdmIncludeAnnotations ParseIncludeAnnotations(IJsonValue jsonValue, IJsonPath jsonPath)
         {
             JsonObjectValue includeObj = jsonValue.ValidateRequiredJsonValue<JsonObjectValue>(jsonPath);
             string termNamespace = null;
@@ -223,7 +303,7 @@ namespace Microsoft.OData.Edm.Csdl.Parsing
 
                     default:
                         // The item objects MAY contain annotations. However, IEdmIncludeAnnotations doesn't support to have annotations.
-                        propertyValue.ReportUnknownMember(jsonPath, options);
+                        propertyValue.ReportUnknownMember(jsonPath, _options);
                         break;
                 }
             });
@@ -231,27 +311,31 @@ namespace Microsoft.OData.Edm.Csdl.Parsing
             return new EdmIncludeAnnotations(termNamespace, qualifier, targetNamespace);
         }
 
-        private static Version GetCsdlVersion(JsonObjectValue csdlObject, JsonPath parentPath)
+        private static TextReader LoadBuiltInVocabulary(Uri uri)
         {
-            Debug.Assert(csdlObject != null);
+            string sourceUri = uri.OriginalString;
+            if (sourceUri.Contains("Org.OData.Core.V1"))
+            {
+                return null;
+            }
+
+            return null;
+        }
+
+        private static Version ParseVersion(IJsonValue jsonValue, IJsonPath parentPath)
+        {
+            Debug.Assert(jsonValue != null);
 
             Version version = null;
-            IJsonValue versionValue;
-            if (csdlObject.TryGetValue("$Version", out versionValue))
+
+            string strVersion = jsonValue.ParseAsStringPrimitive(parentPath);
+            if (strVersion == "4.0")
             {
-                parentPath.Push("$Version");
-
-                string strVersion = versionValue.ParseAsStringPrimitive(parentPath);
-                if (strVersion == "4.0")
-                {
-                    version = EdmConstants.EdmVersion4;
-                }
-                else if (strVersion == "4.01")
-                {
-                    version = EdmConstants.EdmVersion401;
-                }
-
-                parentPath.Pop();
+                version = EdmConstants.EdmVersion4;
+            }
+            else if (strVersion == "4.01")
+            {
+                version = EdmConstants.EdmVersion401;
             }
 
             if (version != null)
