+++ src/Microsoft.OData.Edm/Csdl/Json/Reader/IJsonReader.cs
@@ -4,7 +4,7 @@
 // </copyright>
 //---------------------------------------------------------------------
 
-namespace Microsoft.OData.Edm.Csdl.Json
+namespace Microsoft.OData.Edm.Csdl.Json.Reader
 {
     /// <summary>
     /// Interface for a class that can read arbitrary JSON.
