+++ src/Microsoft.OData.Edm/Csdl/Json/Value/IJsonValue.cs
@@ -4,7 +4,7 @@
 // </copyright>
 //---------------------------------------------------------------------
 
-namespace Microsoft.OData.Edm.Csdl.Json
+namespace Microsoft.OData.Edm.Csdl.Json.Value
 {
     /// <summary>
     /// Defines JSON Value types.
