+++ src/Microsoft.OData.Edm/Csdl/Json/Value/JsonPrimitiveValue.cs
@@ -4,7 +4,7 @@
 // </copyright>
 //---------------------------------------------------------------------
 
-namespace Microsoft.OData.Edm.Csdl.Json
+namespace Microsoft.OData.Edm.Csdl.Json.Value
 {
     /// <summary>
     /// Represents a JSON primitive value.
