+++ src/Microsoft.OData.Edm/Csdl/Parsing/Ast/CsdlRecordExpression.cs
@@ -4,7 +4,6 @@
 // </copyright>
 //---------------------------------------------------------------------
 
-using Microsoft.OData.Edm.Csdl.Json;
 using System.Collections.Generic;
 
 namespace Microsoft.OData.Edm.Csdl.Parsing.Ast
@@ -38,28 +37,5 @@ namespace Microsoft.OData.Edm.Csdl.Parsing.Ast
         {
             get { return this.propertyValues; }
         }
-    }
-
-    /// <summary>
-    /// Represents a CSDL Record expression.
-    /// </summary>
-    internal class CsdlJsonValueExpression : CsdlExpressionBase
-    {
-        public CsdlJsonValueExpression(string termName, IJsonValue jsonValue, CsdlLocation location)
-            : base(location)
-        {
-            TermName = termName;
-            JsonValue = jsonValue;
-        }
-
-        public override EdmExpressionKind ExpressionKind
-        {
-            get { return EdmExpressionKind.None; }
-        }
-
-        // Qualified term name
-        public string TermName { get; }
-
-        public IJsonValue JsonValue { get; }
     }
 }
