+++ src/Microsoft.OData.Edm/Microsoft.OData.Edm.csproj
@@ -54,18 +54,34 @@
     <Compile Include="Csdl\CsdlParseException.cs" />
     <Compile Include="Csdl\EdmValueParser.cs" />
     <Compile Include="Csdl\EdmValueWriter.cs" />
-    <Compile Include="Csdl\Json\IJsonValue.cs" />
-    <Compile Include="Csdl\Json\JsonObjectValue.cs" />
-    <Compile Include="Csdl\Json\JsonArrayValue.cs" />
-    <Compile Include="Csdl\Json\JsonPrimitiveValue.cs" />
-    <Compile Include="Csdl\Json\IJsonReaderExtensions.cs" />
-    <Compile Include="Csdl\Json\IJsonValueExtensions.cs" />
-    <Compile Include="Csdl\Json\IJsonReader.cs" />
-    <Compile Include="Csdl\Json\JsonReader.cs" />
-    <Compile Include="Csdl\Json\JsonReaderOptions.cs" />
-    <Compile Include="Csdl\Json\JsonConstants.cs" />
-    <Compile Include="Csdl\Json\JsonNodeType.cs" />
+    <Compile Include="Csdl\Json\Ast\CsdlJsonModel.cs" />
+    <Compile Include="Csdl\Json\Ast\CsdlJsonSchema.cs" />
+    <Compile Include="Csdl\Json\Ast\CsdlJsonSchemaActionItem.cs" />
+    <Compile Include="Csdl\Json\Ast\CsdlJsonSchemaComplexItem.cs" />
+    <Compile Include="Csdl\Json\Ast\CsdlJsonSchemaEntityContainerItem.cs" />
+    <Compile Include="Csdl\Json\Ast\CsdlJsonSchemaEntityItem.cs" />
+    <Compile Include="Csdl\Json\Ast\CsdlJsonSchemaEnumItem.cs" />
+    <Compile Include="Csdl\Json\Ast\CsdlJsonSchemaFunctionItem.cs" />
+    <Compile Include="Csdl\Json\Ast\CsdlJsonSchemaItem.cs" />
+    <Compile Include="Csdl\Json\Ast\CsdlJsonSchemaOperationItem.cs" />
+    <Compile Include="Csdl\Json\Ast\CsdlJsonSchemaStructuredItem.cs" />
+    <Compile Include="Csdl\Json\Ast\CsdlJsonSchemaTermItem.cs" />
+    <Compile Include="Csdl\Json\Ast\CsdlJsonSchemaTypeDefinitionItem.cs" />
+    <Compile Include="Csdl\Json\Parser\CsdlJsonModelParser.cs" />
+    <Compile Include="Csdl\Json\Parser\CsdlJsonSchemaParser.cs" />
+    <Compile Include="Csdl\Json\Value\IJsonValue.cs" />
+    <Compile Include="Csdl\Json\Value\JsonObjectValue.cs" />
+    <Compile Include="Csdl\Json\Value\JsonArrayValue.cs" />
+    <Compile Include="Csdl\Json\Value\JsonPrimitiveValue.cs" />
+    <Compile Include="Csdl\Json\Value\IJsonValueExtensions.cs" />
+    <Compile Include="Csdl\Json\Reader\IJsonReaderExtensions.cs" />
+    <Compile Include="Csdl\Json\Reader\IJsonReader.cs" />
+    <Compile Include="Csdl\Json\Reader\JsonReader.cs" />
+    <Compile Include="Csdl\Json\Reader\JsonReaderOptions.cs" />
+    <Compile Include="Csdl\Json\Reader\JsonConstants.cs" />
+    <Compile Include="Csdl\Json\Reader\JsonNodeType.cs" />
     <Compile Include="Csdl\Json\JsonPath.cs" />
+    <Compile Include="Csdl\Json\JsonValueParseExtensions.cs" />
     <Compile Include="Csdl\Reader\SchemaJsonItemParser.cs" />
     <Compile Include="Csdl\Reader\AnnotationJsonBuilder.cs" />
     <Compile Include="Csdl\Reader\EdmJsonBuilder.cs" />
