+++ test/FunctionalTests/Microsoft.OData.Edm.Tests/Csdl/Json/JsonObjectValueTests.cs
@@ -4,7 +4,7 @@
 // </copyright>
 //---------------------------------------------------------------------
 
-using Microsoft.OData.Edm.Csdl.Json;
+using Microsoft.OData.Edm.Csdl.Json.Value;
 using Xunit;
 
 namespace Microsoft.OData.Edm.Tests.Csdl.Json
