+++ test/FunctionalTests/Microsoft.OData.Edm.Tests/Csdl/Parsing/CsdlJsonParserTests.cs
@@ -10,6 +10,10 @@ using System.IO;
 using System.Linq;
 using Microsoft.OData.Edm.Csdl;
 using Microsoft.OData.Edm.Csdl.Json;
+using Microsoft.OData.Edm.Csdl.Json.Ast;
+using Microsoft.OData.Edm.Csdl.Json.Parser;
+using Microsoft.OData.Edm.Csdl.Json.Reader;
+using Microsoft.OData.Edm.Csdl.Json.Value;
 using Microsoft.OData.Edm.Csdl.Parsing;
 using Microsoft.OData.Edm.Csdl.Reader;
 using Microsoft.OData.Edm.Validation;
